requirements.txt
# Core Framework
streamlit>=1.29.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.1
langchain-community>=0.0.10

# Google Generative AI (compatible with langchain-google-genai 1.0.1)
google-generativeai>=0.4.1,<0.5.0

# Apify
apify-client>=1.5.0

# HTTP & Web Scraping
requests>=2.31.0
httpx>=0.25.2
beautifulsoup4>=4.12.2

# Data Processing
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.9.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Embeddings (for semantic matching) - Python 3.10 compatible versions
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
numpy>=1.23.0,<1.25.0